# api/openapi.yaml
openapi: 3.0.3
info:
  title: blowball Multi-Agent Backend API
  version: "1.1.0"
  description: |
    HTTP API surface for the blowball multi-agent backend.

    Authentication is JWT Bearer. All routes except `/auth/login` require an
    `Authorization: Bearer <token>` header. Errors follow the unified envelope
    `{"error": {"code": "ERROR_CODE", "message": "..."}}`.

    1.1.0 — turn-detach-resume (disconnecting from an SSE stream no longer
    cancels the turn; run ids, session-busy 409, cancel/resume-by-run-id
    endpoints, `generating`/`run_id` on the session list) and
    per-request-model-selection (`model`/`reasoning_effort` request params,
    `GET /api/v1/models`).

    Workspace file endpoints are transparent to the physical storage backend:
    requests and responses are identical whether `storage.workspace.backend` is
    `local` (per-process local disk) or `shared` (an operator-mounted shared
    POSIX filesystem such as MinIO-backed JuiceFS). The backend selection only
    changes where files physically live; it does not alter this contract.
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token issued by POST /api/v1/auth/login.
  schemas:
...
    LoginRequest:
      type: object
      required: [username]
      properties:
        username:
          type: string
        password:
          type: string
          description: >-
            User password, verified against the seed-created bcrypt hash.
            Required only when auth.password_required is true (the default);
            in passwordless mode the password, if sent, is ignored.
    LoginResponse:
      type: object
      required: [user_id, access_token, expire, token_type]
      properties:
        user_id:
          type: string
          description: The authenticated user's unique ID.
        access_token:
          type: string
          description: Signed JWT to send as `Bearer` in subsequent requests.
        expire:
          type: integer
          format: int64
          description: Token expiry as a Unix timestamp (seconds).
        token_type:
          type: string
          enum: [Bearer]
          example: Bearer
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Exchange a username for a JWT access token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
            example: {username: alice}
      responses:
        '200':
          description: Successful login.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LoginResponse'}
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: |
            Authentication failed. The `message` distinguishes the cause:
            `user not found` or `user disabled`. Unexpected internal failures
            collapse to a generic `unauthorized`.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
              examples:
                userNotFound:    {value: {error: {code: UNAUTHORIZED, message: "user not found"}}}
                userDisabled:    {value: {error: {code: UNAUTHORIZED, message: "user disabled"}}}
                unauthorized:    {value: {error: {code: UNAUTHORIZED, message: "unauthorized"}}}
  /api/v1/sessions:
    get:
      tags: [Session]
      summary: List the authenticated user's sessions, newest first.
      responses:
        '200':
          description: Session list (empty array when none exist).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SessionListResponse'}
              examples:
                populated:
                  value:
                    sessions:
                      # A session with a turn still running (turn-detach-resume):
                      # generating=true carries the active run_id — the
                      # attach/cancel target and the reload recovery path.
                      - session_id: sess_1
                        title: "Plan release"
                        update_time: "2026-06-11T10:00:00Z"
                        generating: true
                        run_id: "018f9b8a-7c1a-7e6a-8e6a-6c1a7c1a7c1a"
                      # An idle session: generating=false, run_id omitted.
                      - session_id: sess_2
                        title: "Notes"
                        update_time: "2026-06-10T09:00:00Z"
                        generating: false
                empty: {value: {sessions: []}}
        '500':
          $ref: '#/components/responses/Internal'
    post:
      tags: [Session]
      summary: Create a new session.
      description: |
        The server mints a UUID v7 session_id, persists the session row, and
        returns it. Clients must create a session before sending messages.
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CreateSessionResponse'}
        '500':
          $ref: '#/components/responses/Internal'
  /api/v1/sessions/{session_id}:
    delete:
      tags: [Session]
      summary: Delete a session owned by the authenticated user.
      description: |
        Archives the session and all of its titles/messages verbatim into the
        `*_deleted` mirror tables, purges the live rows (archive + purge is a
        single atomic transaction), and removes the warm-tier session file.
        Redis cache entries (`session:{id}`, `msgs:{id}`) are intentionally not
        cleared; every read path re-validates ownership against MySQL first, so
        stale cache keys are unreachable until they expire on TTL. Deleting a
        session that does not exist or belongs to another user returns 404; the
        existence of another user's session is never disclosed.
      parameters:
        - name: session_id
          in: path
          required: true
          schema: {type: string}
          description: ID of the session to delete.
      responses:
        '204':
          description: Session deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
    patch:
      tags: [Session]
      summary: Update the title of a session owned by the authenticated user.
      description: |
        Sets a manual title for the session. Manual titles are not overwritten
        by asynchronous AI title generation and the session's update_time is
        refreshed so the session appears first in the list.
      parameters:
        - name: session_id
          in: path
          required: true
          schema: {type: string}
          description: ID of the session to update.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateTitleRequest'}
      responses:
        '200':
          description: Title updated successfully.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UpdateTitleResponse'}
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
  /api/v1/sessions/{session_id}/messages:
    get:
      tags: [Session]
      summary: List the messages of a session with cursor pagination.
      description: |
        Returns the full event stream of a session owned by the authenticated
        user. Messages are ordered by `(msg_time, msg_index)` ascending by
        default; use `order=desc` to retrieve newest first. Pass
        `page_token` from the previous response to advance pages.
      parameters:
        - name: session_id
          in: path
          required: true
          schema: {type: string}
        - name: page_token
          in: query
          required: false
          schema: {type: string}
          description: Opaque cursor from the previous page's `next_page_token`.
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            default: 50
            maximum: 200
          description: Number of messages per page. Values above 200 are clamped.
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort order for `(msg_time, msg_index)`.
      responses:
        '200':
          description: Paginated message list.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SessionMessagesResponse'}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
    post:
      tags: [Session]
      summary: Send a user message and stream the agent's response as SSE.
      description: |
        The session must already exist and belong to the authenticated user.
        Sending to a non-existent session or another user's session returns 404.
      parameters:
        - name: session_id
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SendMessageRequest'}
      responses:
        '200':
          description: |
            Server-Sent Events stream. Each event frame is `id: <cursor>`,
            `event: <type>`, and `data: <json>` followed by a blank line
            (turn-detach-resume: the `id` line carries the run event log's
            stream entry id; native EventSource clients replay it as
            `Last-Event-ID` when resuming via
            `GET /api/v1/sessions/{session_id}/turns/{run_id}/events`).
            The response carries an `X-Run-Id` header (the run id, equal to
            the request's trace id); every event's `meta.run_id` carries the
            same value. Disconnecting from this stream does NOT cancel the
            turn — cancel explicitly via the cancel endpoint.
            Event types: `agent_start`, `token`, `reasoning`, `tool_call`, `tool_result`,
            `agent_end`, `agent_error`, `done`. The `done` event is always the final event.
          headers:
            X-Run-Id:
              description: The run id of this turn (equals the request's trace id) — the cancel/resume target.
              schema: {type: string}
          content:
            text/event-stream:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SSEAgentStart'
                  - $ref: '#/components/schemas/SSEToken'
                  - $ref: '#/components/schemas/SSEReasoning'
                  - $ref: '#/components/schemas/SSEToolCall'
                  - $ref: '#/components/schemas/SSEToolResult'
                  - $ref: '#/components/schemas/SSEAgentEnd'
                  - $ref: '#/components/schemas/SSEAgentError'
                  - $ref: '#/components/schemas/SSEDone'
        '400':
          description: |
            Malformed body, or a rejected per-request model selection
            (per-request-model-selection): `INVALID_MODEL` (unknown catalog
            name) or `INVALID_EFFORT` (invalid effort value, non-`none`
            effort explicitly requested on a `thinking: false` entry, or an
            effort conflicting with a configured agent `output_schema`).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: |
            The session already has a running turn (single active run per
            session). The error object carries the active `run_id` so the
            client can attach to the running turn instead of retrying.
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error:
                    type: object
                    required: [code, message, run_id]
                    properties:
                      code:
                        type: string
                        enum: [SESSION_BUSY]
                      message:
                        type: string
                      run_id:
                        type: string
                        description: The currently running turn's run id (attach target).
        '500':
          $ref: '#/components/responses/Internal'
  /api/v1/sessions/{session_id}/turns/{run_id}/cancel:
    post:
      tags: [Session]
      summary: Cancel a running turn by run id (turn-detach-resume).
      description: |
        Cancels the turn identified by `run_id` and releases its resources.
        Three outcomes: the run is in THIS agent process (cancelled
        immediately, partial output persists through the interrupted-turn
        path), the run heartbeats on another replica (a Redis cancel flag is
        raised; the owning process cancels within one heartbeat period), or
        the run is dead (crashed process — force-cleared to `interrupted`,
        session unlocked). Cancelling an already-terminal run is idempotent
        and reports the terminal status. The run must belong to the caller
        and the path session, otherwise 404.
      parameters:
        - name: session_id
          in: path
          required: true
          schema: {type: string}
        - name: run_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Cancellation accepted (or already terminal).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TurnStatusResponse'}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
  /api/v1/sessions/{session_id}/turns/{run_id}/events:
    get:
      tags: [Session]
      summary: Resume a run's event stream over SSE (turn-detach-resume).
      description: |
        Replays the run's event log from the beginning (or strictly after the
        `Last-Event-ID` request header) and then tails live output until the
        run reaches a terminal state — the endpoint a client uses when
        reopening a session whose previous request is still generating.
        Multiple concurrent subscribers (multiple tabs) are independent. A
        terminal-but-still-retained run replays fully and closes; a run whose
        retention window passed (or never existed) returns 410 — fall back to
        the ordinary message-history read. If a status-running run's process
        died, the stream replays what was logged, then emits a synthesized
        terminal `done` event carrying `usage.error = "interrupted..."`.
      parameters:
        - name: session_id
          in: path
          required: true
          schema: {type: string}
        - name: run_id
          in: path
          required: true
          schema: {type: string}
        - name: Last-Event-ID
          in: header
          required: false
          schema: {type: string}
          description: Stream entry id from a prior subscription's last received frame; resume strictly after it.
      responses:
        '200':
          description: |
            Server-Sent Events stream of the run's events (same frame format
            as POST /messages, including `id:` lines and `X-Run-Id` header).
            The stream ends after the terminal `done` event.
          headers:
            X-Run-Id:
              description: The run id being replayed/tailed.
              schema: {type: string}
          content:
            text/event-stream:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SSEAgentStart'
                  - $ref: '#/components/schemas/SSEToken'
                  - $ref: '#/components/schemas/SSEAgentError'
                  - $ref: '#/components/schemas/SSEDone'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: The run is no longer replayable (retention window passed or unknown); read the ordinary message history instead.
        '500':
          $ref: '#/components/responses/Internal'
  /api/v1/workspace/files:
    get:
      tags: [Workspace]
      summary: List files in (a subdirectory of) the user's workspace.
      parameters:
        - name: path
          in: query
          required: false
          schema: {type: string}
          description: Optional workspace-relative subdirectory. Defaults to the workspace root.
        - name: include_hidden
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Whether to include hidden files and directories (names starting with '.'). Defaults to false.
      responses:
        '200':
          description: File/dir entries, sorted (directories first, then name).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FileListResponse'}
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/Internal'
...
  /api/v1/skills:
    get:
      tags: [Skills]
      summary: List the authenticated user's skills.
      responses:
        '200':
          description: Skill list (empty array when none exist).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SkillsResponse'}
        '500':
          $ref: '#/components/responses/Internal'
...